+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.8.2] - 2026-08-07
 
 ### Changed
@@ -213,6 +215,7 @@
 
 Initial public release.
 
+[unreleased]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.2...HEAD
 [0.8.2]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/mattiejastje/HitmanTracker/compare/v0.7.0...v0.8.0