+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.8.3] - 2026-08-18
 
 ### Added
 
@@ -245,7 +245,7 @@
 
 Initial public release.
 
-[unreleased]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.2...HEAD
+[0.8.3]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/mattiejastje/HitmanTracker/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/mattiejastje/HitmanTracker/compare/v0.7.0...v0.8.0